Add images and create Card, FormField and Loader components, and Create layout of Home page
This commit adds two images to the assets folder and creates Card.jsx, FormField.jsx and Loader.jsx components in the components folder. It also creates a constants folder and adds a surpriseMePrompts export to index.js.

The Home.jsx file in the pages folder is updated to import and use the new components, Additionally, state variables for loading, allPost and searchText are defined and used in the RenderCards function to conditionally render a list of Card components based on the data passed as a prop.