Refactor code and add utility function to generate random prompts
In this commit, a new 'utils' folder has been created and an 'index.js' file has been added. The file exports a utility function named 'getRandomPrompt' that takes an array of prompts as an argument and returns a randomly selected prompt from the 'surpriseMePrompts' array.

In the 'CreatePost.jsx' file, the utility function has been imported and added to the component's state. This function will be used to generate random prompts for the user.

Changes Made:

Created a new 'utils' folder and added an 'index.js' file
Added a utility function named 'getRandomPrompt' to the 'index.js' file
Imported the 'getRandomPrompt' function to the 'CreatePost.jsx' file and added it to the component's state